Train a transformer on MAJORITY (#2)
* finish writing attention layer

* interptransformer tests pass

* tests pass

* add SingleLayerTransformer

* model converges

* model converges

* model converges

* visualization + eval loop added

* add ds_store to gitignore

* update transformer

* update folder name

+++ src/miexp/bfuncs.py
@@ -14,7 +14,7 @@ class MajDataset(Dataset):
         """
         self.N = N
         self.num_samples = num_samples
-        self.data = torch.randint(0, 2, (self.num_samples, self.N)).type(torch.float)
+        self.data = torch.randint(0, 2, (self.num_samples, self.N))
         self.labels = (torch.sum(self.data, dim=1) > (self.N // 2)).type(torch.int)
 
     def __len__(self) -> int:  # noqa: D105
